+++ linen-modern/src/main/scala/x7c1/linen/modern/accessor/unread/UnreadChannelAccessor.scala
@@ -2,7 +2,7 @@ package x7c1.linen.modern.accessor.unread
 
 import android.database.Cursor
 import x7c1.linen.database.{LinenOpenHelper, Query}
-import x7c1.linen.modern.accessor.preset.ClientAccount
+import x7c1.linen.domain.account.ClientAccount
 import x7c1.linen.modern.accessor.unread.ChannelAccessorError.UnexpectedError
 import x7c1.linen.modern.accessor.unread.ChannelLoaderEvent.{AccessorError, Done}
 import x7c1.wheat.macros.database.{TypedCursor, TypedFields}
