+++ linen-modern/src/main/scala/x7c1/linen/modern/display/settings/ChannelRowAdapter.scala
@@ -2,7 +2,7 @@ package x7c1.linen.modern.display.settings
 
 import android.support.v7.widget.RecyclerView.Adapter
 import android.view.ViewGroup
-import x7c1.linen.domain.AccountIdentifiable
+import x7c1.linen.domain.account.AccountIdentifiable
 import x7c1.linen.glue.res.layout.{SettingMyChannelRow, SettingMyChannelRowFooter, SettingMyChannelRowItem}
 import x7c1.linen.modern.accessor.setting.{MyChannel, MyChannelFooter, MyChannelRow}
 import x7c1.wheat.lore.resource.AdapterDelegatee
