+++ linen-modern/src/main/scala/x7c1/linen/modern/display/settings/SourceMenuSelected.scala
@@ -1,8 +1,8 @@
 package x7c1.linen.modern.display.settings
 
 import android.view.View
+import x7c1.linen.domain.account.ClientAccount
 import x7c1.linen.modern.accessor.SettingSource
-import x7c1.linen.modern.accessor.preset.ClientAccount
 
 class SourceMenuSelected private (
   val targetView: View,
