+++ linen-modern/src/main/scala/x7c1/linen/modern/display/settings/SourceRowAdapter.scala
@@ -5,9 +5,9 @@ import android.support.v7.widget.RecyclerView.Adapter
 import android.view.{View, ViewGroup}
 import android.widget.SeekBar.OnSeekBarChangeListener
 import android.widget.{RelativeLayout, SeekBar}
+import x7c1.linen.domain.account.ClientAccount
 import x7c1.linen.glue.res.layout.SettingChannelSourcesRow
 import x7c1.linen.modern.accessor.SettingSourceAccessor
-import x7c1.linen.modern.accessor.preset.ClientAccount
 import x7c1.wheat.ancient.resource.ViewHolderProvider
 import x7c1.wheat.modern.decorator.Imports._
 import x7c1.wheat.modern.resource.MetricsConverter
