+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/PresetsAllFragment.scala
@@ -4,7 +4,7 @@ import android.content.Context
 import android.os.Bundle
 import android.view.{LayoutInflater, View, ViewGroup}
 import x7c1.linen.database.{ChannelSubscriber, LinenOpenHelper}
-import x7c1.linen.domain.AccountIdentifiable
+import x7c1.linen.domain.account.AccountIdentifiable
 import x7c1.linen.glue.res.layout.{SettingPresetChannelRow, SettingPresetTabAll}
 import x7c1.linen.modern.accessor.setting.AllPresetChannelsAccessor
 import x7c1.wheat.ancient.resource.ViewHolderProviderFactory
