+++ linen-modern/src/main/scala/x7c1/linen/modern/init/unread/UnreadChannelsReader.scala
@@ -1,6 +1,6 @@
 package x7c1.linen.modern.init.unread
 
-import x7c1.linen.modern.accessor.preset.ClientAccount
+import x7c1.linen.domain.account.ClientAccount
 import x7c1.linen.modern.accessor.unread.ChannelLoaderEvent.Done
 import x7c1.linen.modern.accessor.unread.ChannelSelectable
 import x7c1.wheat.macros.logger.Log
