--- linen-repository/src/main/scala/x7c1/linen/domain/AccountIdentifiable.scala
@@ -1,11 +1,0 @@
-package x7c1.linen.domain
-
-object AccountIdentifiable {
-  def apply(id: Long): AccountIdentifiable = new AccountIdentifiable {
-    override def accountId: Long = id
-  }
-}
-
-trait AccountIdentifiable {
-  def accountId: Long
-}
